refactor code according to master change
